split publishable code from test code
