+++ mps-json-exporter/.gitignore
@@ -1,1 +1,1 @@
-/gen
+/src-gen
