+++ mps-json-exporter/src-test/jsonSchema/test.ts
@@ -1,12 +1,13 @@
 import {generate} from "./generator.ts"
-import {asPrettyString} from "../utils.ts"
-import {JsonSchema} from "../../gen/type-defs/JsonSchema-types.ts"
+import {asPrettyString} from "../../src/utils.ts"
+import {JsonSchema} from "../../src-gen/type-defs/JsonSchema-types.ts"
 import {loadModel} from "../dev-model-loader.ts"
+import {ensureDirSync} from "https://deno.land/std/fs/mod.ts";
 
 
 /*
  * run on CLI:
-    deno run --allow-read --allow-write src/jsonSchema/test.ts
+    deno run --allow-read --allow-write src-test/jsonSchema/test.ts
  */
 
 
@@ -16,7 +17,8 @@ const schemas = loadModel<JsonSchema>(
 )
 
 
-const basePath = "./gen/jsonSchema-examples/"
+const basePath = "./src-gen/jsonSchema-examples/"
+ensureDirSync(basePath)
 
 // generate a separate JSON (Schema) file for every {@link JsonSchema} instance in the model:
 schemas.forEach((schema) => {
