+++ mps-json-exporter/src-test/languageStructure/structure-types.ts
@@ -4,7 +4,7 @@
  */
 
 
-import {Named, Node, Reference} from "../generic.ts"
+import {Named, Node, Reference} from "../../src/generic.ts"
 
 
 export interface EnumerationMemberDeclaration extends Node, Named {
